pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "support-triage-openenv"
version = "0.1.0"
description = "A real-world OpenEnv environment for SaaS support operations triage."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "openenv-core==0.2.3",
  "openai>=2.7.2,<3",
  "fastapi>=0.115.0,<1",
  "uvicorn>=0.30.0,<1",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools]
include-package-data = true
